requirements.txt
absl-py==2.1.0
altair==5.5.0
asttokens==2.2.1
attrs==25.3.0
audioread==3.0.1
backcall==0.2.0
blinker==1.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.3.3
cycler==0.12.1
Cython==3.0.8
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
immutabledict==4.2.0
ipykernel==6.23.2
ipython==8.14.0
isodate==0.6.1
jedi==0.18.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kagglehub==0.3.12
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.10.5
matplotlib-inline==0.1.6
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.0.7
narwhals==1.41.0
nest-asyncio==1.5.6
networkx==3.2.1
noisereduce==3.0.0
numba==0.59.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
ortools==9.9.3963
packaging==23.1
pandas==2.3.1
parso==0.8.3
patsy==1.0.1
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.5.3
plotly==6.1.2
pooch==1.8.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAutoGUI==0.9.54
pybind11==2.11.1
pycparser==2.21
pydeck==0.9.1
pydub==0.25.1
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
pyngrok==7.2.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyroomacoustics==0.7.3
PyScreeze==0.1.30
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.0
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.16.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.2
soundfile==0.12.1
soxr==0.3.7
stack-data==0.6.2
statsmodels==0.14.4
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.2.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.13.2
tzdata==2024.1
ultralytics==8.3.176
ultralytics-thop==2.0.15
uritemplate==4.1.1
urllib3==2.2.0
watchdog==6.0.0
wcwidth==0.2.6
xgboost==3.0.2
youtube-transcript-api==0.6.2
